requirements.txt
# Core dependencies
pytest==7.4.3
pytest-cov==4.1.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
anthropic==0.5.0
pyyaml==6.0.1